requirements.txt
streamlit
langchain
langchain-community
pypdf
faiss-cpu
sentence-transformers
google-generativeai==0.8.5
gtts
python-dotenv
